face-recognition/requirements.txt
face_recognition==1.3.0
flask==2.0.1
python-slugify==5.0.2